Revert "Remove logic of polling liveness of compiled code."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/b8e8a0e4-e79f-4343-9f35-57e5ab1cc7ed). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ia7528fdf17d5d9a41688ea20ab5289509b7049b5