Revert "Disable write-barrier elimination pass"
This reverts commit 5a3271d7caafefd10a20f5a5db09d2c178838b76.

Reason for revert: PS1 is the plain revert

This CL has two fixes (codegen not doing a null check if a write
barrier is being relied on, and codegen not recomputing
skipping write barriers), regression tests, and a new
runtime check which runs in debug mode to ensure that the
dirty bit is set in skipped write barriers. It also adds
compile time (graph checker) tests to ensure graph consistency.

This patchset updates the WriteBarrierKind to be
{emit being relied on, emit not being relied on, dont emit}.

Test: art/test/testrunner/testrunner.py --host --64 --optimizing -b
Bug: 301833859
Bug: 310755375
Bug: 260843353
Change-Id: I025597e284b2765986e2091538680ee629fb5ae7