Revert "Remove logic of polling liveness of compiled code."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/12b3e41a-cda7-4be7-a215-a8f3147f43b3). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ia2446a90e5fde2b9510cd55ac635f60cc708082b