Revert "Remove logic of polling liveness of compiled code."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/8c3d2e26-c9c3-415a-b0ea-f50eeac58628). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I33b7c39925ea6db9374622eae70e647023a21237