Revert "Remove logic of polling liveness of compiled code."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/89b399cd-cac1-498e-a33f-a2f5363dabb1). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I79fde4d0d32387b2a097cc93b38a03c2cbb14aa8