Revert "Move the construction of ProfilingInfo in the compiler."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/8a70835f-2ee6-488e-b228-043f365a3db6). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: If560659a650a67f70434425f3510ed40374e4bf3