Revert "Move the construction of ProfilingInfo in the compiler."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/e1896e6f-954e-4dfa-bd9c-55d2d274216c). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I31563ded125950a194d36600910fd054ba644b4e