Revert "Move the construction of ProfilingInfo in the compiler."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/5693ab19-789b-4282-ae8b-d7742b7f4a76). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I4b5eecde259c81d7fa061479e778fb58fd07d5ac