Revert "Revert "Move the construction of ProfilingInfo in the co..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/5e0ff072-94b6-4360-92fe-44ac8bf143cc). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ifcb7428381933eeecf9a8daf5f1e8de0454adb77