Revert "Revert "Move the construction of ProfilingInfo in the co..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/36ce7848-c020-4e7b-9cae-50cee7c019ce). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ibaa200913a751446e67af24966e83093927e3701