Revert "Clean up BUILD_BROKEN_DISABLE_BAZEL that no longer does ..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/bece68ca-d02c-406c-a29e-1910a3315375). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I22d201de3695ef335e8f5a58023a6d776ec424ab