Add lock protection for request synchronous checkpoint of collect lock-holding thread's stack.
When there are classes in the lock-holding thread's execution stack that have not been fully loaded,
and multiple checkpoints have been installed by other waiting for lock threads at this time,
it will cause a deadlock problem in the lock-holding thread.

In order to fix this deadlock problem, we add some synchronization protection to ensure there is only
one checkpoint of collect thread stack can be installed to the lock-holding thread at the same time.

Bug: 317315406

Change-Id: I9625c33bff3a6562d835dd9e5ad809b0ea37ec5e
Signed-off-by: yanmin5 <yanmin5@xiaomi.corp-partner.google.com>