Move code to a helper class ArtFileManager.
This makes ArtManagerLocal.java shorter and enables code reuse, to
prepare for the subsequent change.

Also, this CL also changes PrimaryDexUtils to only return dex info items
that have code except for finding a split specified by the user. This is
done because dex info items that don't have code are generally not
useful to ART Service.

This CL is a no-op change. Existing tests are still passing.

Bug: 317081212
Test: atest ArtServiceTests
Change-Id: Ibce2d62bd5c0311fb4d7f92098dcd3cf38f5984f