Revert "Avoid dynamic lib dependencies on internal ART libraries..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/6454bfbf-a3cf-42b0-9d4e-7fc1b34eb76d). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I5f0ff773cefe5c07db7132dbacdcf86adc75d958