Revert "Avoid dynamic lib dependencies on internal ART libraries..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/7dbd12a4-dc71-4bba-80a1-554bea5ff66d). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I5ee4376cf0c77a3b575412ccb681bd1d4b2e624c