Revert "Avoid dynamic lib dependencies on internal ART libraries..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/fe6f849a-ace7-441e-b328-0fbe33139af7). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I926bab060fc985a5da0511b1d7dd3737b720479e