Revert "Avoid dynamic lib dependencies on internal ART libraries..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/f1077b93-7d52-4c26-98f5-13d51c8e217b). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ie2520fd5da3091dc8ee037249664287bc9458eb5