Revert "Avoid dynamic lib dependencies on internal ART libraries..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/8a8c4a5f-bfa9-4f07-a374-f220946f52ff). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ied53786815dd4b610f670600b600c6653990eeff