Revert "Avoid dynamic lib dependencies on internal ART libraries..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/e94e8dea-6310-49a3-87c2-a2c768a92786). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I8d88b238a74f222a5bb4ea7ef297e19ef5df8ea8