Revert "Run optimizations with baseline compilation."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/eaa2fcdb-8708-44e1-b43e-65eae246a581). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I9932c432c4d2b6cb11a8b99446193ae4ce470efe