Revert "Run optimizations with baseline compilation."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/a64f39cc-486a-47ea-a8d0-ad36fac4c254). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I344fa2e45ca5a0fa747bfe6d1a5e922d0a6125fc