Revert "Run optimizations with baseline compilation."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/345dd777-a6ee-49b4-9dcb-4479af8b7ba6). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I78131a27097372ca32f8c1090fdba447fb9761ce