Revert "Run optimizations with baseline compilation."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/2c1d0c38-701b-4101-ac77-05fb51f4acf5). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I0fcdd37ff4d41bafe2adb169bf979cf26eeecbb7