Make entrypoints and Context ISA configurable
Make the RuntimeContextType and QuickArgumentVisitor ISA configurable
via templates instead of using #ifdef's. This enables the context type
(RuntimeContextType) to be configured via a constexpr which can
support targets (e.g: a simulator) where the quick code ISA is
different to the native ISA as reported by predefined macros (e.g:
__aarch64__, __x86_64__, etc...).

Code size is not impacted as the constexpr only allows one context to
be used at a time thus only one context will be included in the
binary.

Author: Artem Serov <artem.serov@linaro.org>
Author: Chris Jones <christopher.jones@arm.com>

Test: test.py --host --target
Change-Id: I442d3686a6af51268e118f2d0e408f16ae1faca0