[Sim] Support kCallOnMainOnly in restricted mode
Support compilation (and hence simulation) of methods that have
kCallOnMainOnly in their LocationSummary. This enables simulation of
HThrow, HMonitorOperation, HNewArray, HNewInstance and HRem HIRs.

To support this change in the simulator, many more C++ quick
entrypoints are now set to be intercepted by VIXL.

In addition because ldxr and stxr are not properly simulated, always
use the slow locking/unlocking path to enable HMonitorOperation to
be simulated.

Note: the compilation filter has been removed which allows
kCallOnMainAndSlowPath instructions to be compiled although in
practice this only applies to intrinsics which are still disabled in
restricted mode.

Author: Chris Jones <christopher.jones@arm.com>

Test: Follow instructions in README.simulator.md
Test: ./test.py --host --target --simulator
Test: 530-checker-lse  (HThrow)
Test: 130-hprof        (HNewArray)
Test: 426-monitor      (HMonitorOperation)
Change-Id: I3cc77cbbe98d1ccfecce5e0e1125762c39b8cdea