Revert "Reland "Run optimizations with baseline compilation.""
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/e6e10312-a926-4c6f-8a34-60fde998f15a). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I1de3ef3e107399e2f54df78d66d1bc623ade5fc8