Revert "Run optimizations with baseline compilation."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/36d4b9f7-b6af-4227-aaf1-5464fb056acc). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ia6671bab580edf9ee166d14f8ec16c6d65f7a62a