Revert "Revert "Reland "Run optimizations with baseline compilat..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/e5909544-8981-4641-875c-70a63fd665fa). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Id956cc2d0fb0c68aed2fe9129480adf2ebca360e