Revert "Revert "Reland "Run optimizations with baseline compilat..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/2e84f314-d7f7-4f07-8a7f-abf8df0ff9a7). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: If0fe98f0dad9f747a5e9ce47fad53cd2b8090347