Revert "Revert "Reland "Run optimizations with baseline compilat..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/9108f4eb-2814-4334-9e47-2d75ca5b2902). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I7be2ddf013ebc4cfc23ba3974a7364aa9573f512