Revert "Revert "Reland "Run optimizations with baseline compilat..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/0d43b166-5eb5-42f6-8894-c1e2e91d6ab5). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I689fd4134ae6337290c1860a869b1480316b9056