Revert "Revert "Reland "Run optimizations with baseline compilat..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/c56d4bf2-c252-46b7-9617-023e79269022). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Iee3ddf32d14126dd6dae5c835a8a733a4d586440