Revert "Add visibility attributes in runtime/jni"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/b6235268-b540-4d66-9e64-c0d8a00b9a7d). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ib0bd452f42ec325b9a1692b352eaa7fb95ac6b68