Revert "Add visibility attributes in runtime/jni"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/98596fe6-75f5-4cd8-a68e-e9bd690be257). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I1bb752d55a171c9ed7753a737cc4ca605a158746