Revert "Add visibility attributes in runtime/jni"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/ae1a0afd-8853-4b29-b2df-cbfd191b64f1). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I45dd4de70bf7c150730efeec2a1acf021470d43a