Revert "Fix a class reference."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/cc2f4cdc-f35a-4229-a265-128f8bb74e4f). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I7898dd8de304a03d92a8272531956250e65c90c4