Give full access to native libs from java libs in the same partition (reland).
For both packages and shared java libs in system image partitions
(system, product, vendor), first try to load native libraries from the
same partition, by using the linker namespace for that partition
("default" or "system", "product", "sphal", respectively).

Continue as usual if that fails, i.e. load with the classloader
namespace for the package, which for shared java libs may mean loading
from a different partition.

This change means private native libraries in the same partition now
will take precedence in case of name conflicts, which may be
incompatible. Hence it's only enabled for SDK level 35 (VIC) and later.

A more principled approach would be to extend the library loading API
to allow callers to specify whether loading should be based on the
their own or the package location. At this point there are no plans for
such an extended API.

This is based on the revert in
commit 453b9fe909c22eb0f01b3072a5213dd46aed6f45, but with a different
solution. It also extends the approach to work for vendor and product
partitions.

Test: atest libnativeloader_e2e_tests \
        libnativeloader_test libnativeloader_lazy_test
Test: libnativeloader_e2e_tests on S, Sv2, T, and U platforms in CI
Bug: 237577392
Change-Id: If8b74503edfa9229b9eada73968b7d7b5c75ca10