Revert "Inverse inputs for some vector operations if the generat..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/67deb65f-cb1e-4d8e-ab86-f23785483768). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ie33282232eb7e6e8c3468a306a3e7d460fc4c6e1