Revert "Revert from reporting pulled atom ArtDeviceStatus"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/1538091e-0a23-4ff9-8ccd-f0e00e7e3bcf). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I76e8ffe303825f7dab25fb92c34b2d282924c6c9