[NOT FOR MERGING] Support sparsity in FdFile::Copy
************************************************************************
[NOT FOR MERGING]
Currently there is no production use case for a sparse file copy.
CL raised for future reference when a use-case is identified.
Patchsets:
  * PS2: Userspace sparse copy
  * PS1: FIEMAP + sendfile (kernel-space) sparse copy
************************************************************************

Output a sparse file during FdFile::Copy by reading and checking each
block of the source file in userspace, writing only the non-zero blocks
and skipping over any zeroed blocks (holes). This helps to save disk
space on file systems that support sparse files.

This is based on and replaces a reverted lseek SEEK_DATA/SEEK_HOLE based
sparse file copy (7e06898) following potential issues with lseek
SEEK_DATA on f2fs, where lseek SEEK_DATA will sometimes return an
incorrect offset. This bug is not yet able to be reproduced in the
gtests, but can be reproduced with a manual odrefresh invocation.

The implementation support and tests are refactored over that of 7e06898:

* Add tests for partial sparse-file copying.

* Only dissallow copies to non-empty files if the existing data is in
  the target copy region, and add a test case.

* Add tests that the original source file's sparsity is preserved when
  the starting offset for the copy is aligned to the filesystem
  blocksize. If the starting offset is unaligned, the output file's
  sparsity may differ from that of the source file.

Bug: 323906594

Tested on oriole and shiba (ext4 and f2fs):
Test: dex2oat ... -g --strip -oat-symbols=...
Test: odrefresh --force-compile
Test: art/tools/run-gtests.sh

Co-authored-by: Richard Neill <richard.neill@arm.com>
Co-authored-by: Ruben Ayrapetyan <ruben.ayrapetyan@arm.com>
Change-Id: Idcc5a9c5320b3a80ea5c6186fa4c41d92f1727f8