Preserve file sparsity in FdFile::Copy with FIEMAP
Output a sparse file during FdFile::Copy by using ioctl FIEMAP to
transfer only the allocated filesystem extents and skip over holes.
This helps to save disk space on file systems that support sparse files.

This is based on and replaces a reverted lseek SEEK_DATA/SEEK_HOLE based
sparse file copy (7e06898) following potential issues with lseek
SEEK_DATA on f2fs, where lseek SEEK_DATA will sometimes return an
incorrect offset. This bug is not yet able to be reproduced in the
gtests, but can be reproduced with a manual odrefresh invocation.

The test coverage is also expanded over that of 7e06898:

* Add tests for partial sparse-file copying, which is supported by both
  the previous dense-copy implementation, and the new sparse-copy
  implementation. This also ensures a FIEMAP-specific aspect is tested,
  where copied ranges start/end in the middle of an extent.

* Add a test to copy a larger file with more chunks, to ensure we test
  that the FIEMAP-based algorithm works correctly when more than one
  ioctl call is required (as not all extents can fit in a single 4KB
  buffer).

Some minor changes are made such as using data 'chunk' instead of
'block' to avoid confusion in context of filesystem extents and blocks.

Bug: 323906594

Tested on oriole and shiba (ext4 and f2fs):
Test: odrefresh --force-compile
Test: art/tools/run-gtests.sh

Co-authored-by: Richard Neill <richard.neill@arm.com>
Co-authored-by: Ruben Ayrapetyan <ruben.ayrapetyan@arm.com>
Change-Id: Idcc5a9c5320b3a80ea5c6186fa4c41d92f1727f8