Revert "Remove unused typedefs"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/0a19d592-6d2f-4364-a77c-7f8d320ac64d). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I3d22d5a739f43f83c5c432195be61887cf4cb363