Revert "Increase inlining threshold for baseline."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/72429ecd-cd32-4c54-9054-78caaeb735b0). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I5bd01758ec4f15956a1fb768255ab78a9a63f6e4