Don't lookup for method id when processing trace entries
We used to look up for method id in the method pointer to method id map
to both record method info if this is the first time we are seeing the
method and also to obtain a 32-method id to avoid using 64-bits for
encoding method id. Looking up for the method id on each trace entry is
expensive.

With the new method format we could use 64-bits
since we encode the full 64-bit only once at the start of the block of
trace entries. The subsequent methods are encoded as a diff so using
64-bits doesn't increase the trace file size by a lot. We also record
the methods on class load so we don't need the map to check if we need
to record the method information.

This increases the trace file size but improves the performance. On a
30 seconds trace on google maps, it increases the file size by about 35%
but decreases the time spent on processing trace entries from 3.5 seconds
to 1.2 seconds on main thread and from 12 seconds to 3 seconds on trace
processor thread.

Bug: 259258187
Test: art/test.py
Change-Id: I4f20336624d94793a8f6109086aff55d09dcc54e