Revert "Reduce and consistently report suspend timeouts"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/211318c2-a0de-4b4c-8b24-ae719bfa2b36). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Icc3841e113ed5be3058598280171003686b0652f