Link libnativeloader dynamically into libnativeloader_test again.
It's a CTS test that needs to load libnativeloader dynamically for API
coverage.

Split out the unstable APIs that the test also needs into a separate
library that is linked statically into it. Note that this will still
link some logic statically into tests, like retrieving the lists of
public libraries.

Test: atest libnativeloader_e2e_tests \
        libnativeloader_test libnativeloader_lazy_test
Test: readelf -d out/target/product/vsoc_x86_64/testcases/libnativeloader_test/x86_64/libnativeloader_test64
      -> Check that it only lists NDK libraries.
Test: Remove a symbol from libnativeloader_test.map
      m libnativeloader_test libnativeloader_lazy_test
      -> Check that linking fails.
Bug: 327407752
Change-Id: Ic3fdd0f4c5103c96c303b1f364459059d7ec3044