Revert "Don't set LD_LIBRARY_PATH to the ART module for any tests."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/de842bb0-710d-4494-80ec-ec05fda68948). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ie9f472ad013879b161c50826f0bde880c991eeef