Revert "Don't set LD_LIBRARY_PATH to the ART module for any tests."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/9bf2049b-4d1c-4e3b-9a77-eae537bee686). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Id3153faf97f0b829fc4c061eddc5650af05017a7