16k: Hint that allocation spaces should be 32mb aligned
The CMC GC performs a mremap() spanning the entire bump-pointer space
during a stop-the-world pause. As a result, it's important for the space
to be aligned to at least the Page Middle Directory, allowing mremap()
to copy page tables more efficiently, thanks to having a lot fewer
entries to work on, all of them PMD-sized.

However, when the kernel is configured with 16k pages instead of 4k,
this increases the PMD size from 2m to 32m. This patch adjusts the
allocation hint to be 32m aligned.

Test: m test-art-host-gtest
Test: testrunner.py --host --64 --optimizing -b

Test: run-gtests.sh -- -j6
Test: testrunner.py --target --64 --optimizing

Change-Id: Icf3e5ae38975d35de69cc744750783c50e767029