Revert "libnativebridge_test: link non-public libs statically" and fix the CTS test to only call exported APIs.
CTS tests need to link the libraries they test dynamically, and this
test only tests exported stable APIs. The standalone CTS test config
template includes NativeBridgeModuleController, so the test won't run
in a translated environment (which has no proxy for this library and
hence caused the error that prompted
commit 7f5b3d89f21edd958ddb9d7bd01e03c77bd5425a).

The test also used to test some non-exported APIs, which it shouldn't
do in CTS. This changes it to use art_standalone_test_defaults, which
doesn't have the test_for attributes that made it bypass the stub.

Also reuse the trivial API-only tests for libnativebridge_lazy, and
give them a more apt name for the new use.

Test: atest libnativebridge-tests \
            art_libnativebridge_cts_tests \
            libnativebridge-lazy-tests
Bug: 327407752
Change-Id: Ib56148f1a651a9ffa3045bc401cc625e6f42416b