Add FdFile::Rename
Add the ability to change a file's path via the Rename function, instead
of requiring copy and delete.

On linux, the rename syscall requires the source and destination to be
on the same mounted filesystem. As this approach does not modify the
file's contents, this has the added benefit of ensuring the sparsity of
the resulting file in the destination path is preserved, thereby saving
disk space on systems that support sparse files.

As rename applies changes to parent directory metadata, FdFile now
supports flushing metadata in addition to data (via fsync), as well as
allowing read-only file descriptors to be flushed (required for
directories).

Bug: 323906594

Tested on oriole and shiba (ext4 and f2fs):
Test: odrefresh --force-compile
Test: art/tools/run-gtests.sh

Change-Id: I1d017552fbffcc728cac7a55a14d4d17d1770cb8