Revert "Add `$inline$` to `localStaticNop()` in run-test 569."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/d14ca789-1bd8-42d9-946f-4727af6cda58). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Iefd2171f552d0cc613027ca24395b6bbb6401861