Remove `test_for` from ART tests.
`test_for` bypasses the stubs when tests are built against the shared
libs. However whenever we do that we have to make sure that the tests
only use the stable exported APIs, so the stubs should be used. If the
tests don't find the symbols they need then, they have to link the
library statically.

Test: gtests in ART chroot
Test: Presubmits
Bug: 247108425
Change-Id: I207000d5c9731209aa497e284bbc5106f8ab65d8