Ignore "--staging-dir" passed to odrefresh.
An upcoming change will make odrefresh use `rename` instead of copying
when commiting staging files. This means the staging files must be on
the same filesystem as the output dir. Therefore, we can't respect
"--staging-dir" as it may point to a location on a different filesystem.

Currently, the only user of "--staging-dir" is CompOS. In fact, we don't
need to stage files in CompOS. If the compilation fails (partially or
entirely), CompOS will not sign any artifacts, and odsign will discard
CompOS outputs entirely. After this change, if odrefresh is run in
CompOS, it writes to the output dir directly.

Test: adb shell /apex/com.android.compos/bin/composd_cmd test-compile
Test: -
  1. Install an ART apex.
  2. adb shell cmd jobscheduler run android 5132251
  3. See that CompOS finishes successfully.
  4. Reboot and see that odsign accepts CompOS'es outputs.
Test: atest odsign_e2e_tests_full
Test: atest art_standalone_odrefresh_tests
Change-Id: Icc4c6a378bc85e428003380c7d4014d9b2516c01