Revert "Revert "libnativebridge_test: link non-public libs stati..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/2be76771-ba72-4793-a797-46a184a9bf44). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I67612c683b941417fb8ba9af592fb17d3c7c1b51