Add comment that host variant of libdexfile is dropped in >=V sdks
The host variant of libdexfile and its deps were added to module sdk to
support the use case of building platform tests without art's source
code. This use case is not very relevant today, so we can drop the host
variant of these modules from the module sdk.

We still need to keep them in <= U snapshots, so this CL keeps these
libraries in `native_shared_libs`, and relies on aosp/3011798 to
conditionally drop them from the generated sdk artifacts based on the
value of `SOONG_SDK_SNAPSHOT_TARGET_BUILD_RELEASE`

Test: SOONG_SDK_SNAPSHOT_TARGET_BUILD_RELEASE=UpsideDownCake m
out/soong/mainline-sdks/art-module-host-exports-current.zip # verified
that libartbase is in the generated Android.bp

Test: m out/soong/mainline-sdks/art-module-host-exports-current.zip #
verified that libartbase is not in the generated Android.bp

Bug: 329830257
Test: In `main`, deleted the host variants from prebuilts/module_sdk locally; enabled art prebuilts;
Test: atest --host --no-bazel-mode libunwindstack_unit_test

Change-Id: Ie5b1c25829d3a8f8b8b1f4034532708f8a161216