Revert "Remove unused dex_file field from ParallelCompilationManager"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/3ae5155b-c053-418f-bbb4-eb294e3911a4). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I908530906e372b33d5ebf5fb992703c8f52fed40