Revert "Disable the runtime thread pool."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/b25c1720-2a06-4c83-9717-4f79d21c07ee). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I00d1b3ba9d6cec73f2718dcd6a18e4299d645d4f