Revert "Disable the runtime thread pool."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/e3db9a1d-73eb-42f6-9d88-3c226ec5d760). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I52531a2e527acb6300fcd2ce67577c740952c4b6