Revert "Remove unused dex_file field from ParallelCompilationManager"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/66e49741-3085-4256-a773-af85ab9e8152). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I30e4a914ed4fe0e586e5fba523c276c4bb2df87b