Revert "x86_64: Add JIT support for LoadMethodType."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/699a898e-3c87-472b-806a-8ddf45bef4dd). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I828df31c719895661b3fcaa57894476e2546ffa6