Revert "x86_64: Add JIT support for LoadMethodType."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/efc600e3-97a2-4412-a310-874c8a83ef76). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I68b71a02b712c4b8300eb0d0bfc6bfba8ad20994