CompilationResult: not an aggregate with C++20
As of C++20, the user-declared private CompilationResult() constructor
prevents CompilationResult from being an aggregate class, so aggregate
initialization doesn't work with it anymore.

See https://wg21.link/p1008r1. The paper's authors decided that a
defaulted default constructor should prevent aggregate initialization.
In C++17 and before, the result is surprising:

    CompilationResult do_compile(...) {
      ...
      CompilationResult result;    // ill-formed, calls private ctor
      CompilationResult result {}; // OK, uses aggregate initialization
      ...
    }

In C++20, neither of the above will compile, because
CompilationResult() is private.

Bug: 311052584
Test: remove cpp_std override and build Android
Change-Id: I3ecb8bec5fb69d453d82fff7235caf417c9f1a8c