Revert "Revamp JIT GC."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/8c992847-ebe9-4e8b-aa7d-d63723c1718f). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I5cc5be440f7e1d74f63956e467e4a257f43d0883