Revert "Reland "Revamp JIT GC.""
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/fcdaa2c9-cb6a-46d3-967a-3a19f9add0be). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I711f396fda6b43421448c952f9cda8ac48169a3d