Revert "Apply `CompilerDriver::LoadImageClasses()` to app images."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/e8a38468-fe88-40da-9639-f95b1a5d4cd4). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ic8b0268a5357050c8a75d723bd9e2fa57c292719