Revert "Refactor error reporting in `OatFileBase::Setup()`."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/32e11839-c3d4-45fa-9659-5872d53508d6). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I5dcf5cf61922a9d25eb9b714b0e7763a0fe755a8