Revert "Adjust JIT thresholds."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/8c5f9992-642b-486c-8c98-eea949bb552c). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Iefa9b935a7214d84fe35b6b90e1a8b4a2f005ab5