Revert "Adjust JIT thresholds."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/69d2c07a-b9d7-43a0-9c7f-1a9799bff921). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I6ce0a8ad23112b36bd58ebb9c162f5b208e91312