Revert "Adjust JIT thresholds."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/8298ebb0-b856-4b2b-b4b1-02adc6ea2b13). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I962c8886d217c3ddd3663bccc28993cee8bc8f2d