Revert "Adjust JIT thresholds."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/26ee925d-7b5b-4ae8-a376-8b49c78fe6e7). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I8c09501764a3f3c664ea0f0bfa85cc50f442d674