Revert "Adjust JIT thresholds."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/716f62fd-fd8a-40c8-847b-9e63b57f1cc2). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I023bde7ed9a56a1541dfb5597f80ff53981b7c02