Revert "Adjust JIT thresholds."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/a4d71929-494a-483a-b6a1-bb1058352852). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I101879917f89739632b7748ba09f602a4a24cc85