Revert "Adjust JIT thresholds."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/ccb0a720-5e9e-473b-a477-bf2df8f3dbcc). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: If285b19815e1e4f617aba058c7d99abbcdcc037f