Revert "Reland^2 "Revamp JIT GC.""
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/c3e9f25a-5d48-4063-b6cc-69fa5479a9d3). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: If3db06b53bc76ac219aed1d09a4c47e77b72f19e