Delay string_view construction
Crude tracing suggests the vast majority of the time we leave
Lookup() prior to needing to check against our 'first_checked_str'.
So we delay the construction of our std::string_view, which
incurs a call to strlen(), until we know we'll need it.

On a mid-tier device, this saw strlen() drop from 1.01% to 0.35%
of CPU cycles when installing one app, and from 1.19% to 0.45%
when installing another.

Test: Use simpleperf to check before and after this change, when doing an "adb install-multiple *.apk *.dm" of a couple apps.
Bug: 338123769
Change-Id: I47804769330d47ee0ae19102b9656dfc2937c075