Revert "Adjust JIT thresholds."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/25564fa9-5338-4e54-8efb-b752c398952f). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Idf15bd96d2e2be01fdab6824d50e6282fd02b30a