Revert "Reland^2 "Revamp JIT GC.""
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/72a3b2a7-9619-497b-8850-57e398dddcf8). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I0332d9709482da82b12ebbc05684c4f047ebc283