Revert "Adjust JIT thresholds."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/938cdb8c-440a-44ea-b34e-b3e6a7ac77f8). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I0fbc00c74994e226735700c0192faf111a30b3d6