Revert "Adjust JIT thresholds."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/2e1b7b8c-1e91-4c58-b640-7d2188e4d162). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I6cb4e3ed4eaecdcdf36e1cab886b8a52523c93ae