Revert "Adjust JIT thresholds."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/55d3d7d1-de4b-4459-b67e-dc3fcc0c5c91). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I90ad0edcab0066f4aab27bd453334a36cc66c4d3