Revert "Adjust JIT thresholds."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/e5e8a123-c7d0-42dc-81ad-0f82a25757b1). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I5b3272a4d59229d1caf28ca317cd101164f30587
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com